Unacceptable code from nick. I don't tolerate warnings :3

+++ projects/Engine/src/graphics/vk/VulkanFramebuffer.cpp
@@ -41,7 +41,7 @@ void VulkanFramebuffer::construct(const FramebufferCreateInfo& aInfo)
 	info.height = aInfo.height;
 	info.layers = aInfo.layers;
 	info.renderPass = rp->getHandle();
-	info.attachmentCount = attachments.size();
+	info.attachmentCount = static_cast<uint32_t>(attachments.size());
 	info.pAttachments = attachments.data();
 
 	const VkResult res = vkCreateFramebuffer(context->getDevice(), &info, nullptr, &mBuffer);
